requirements.txt
#requirements.txt
#venv\Scripts\activate
#pip install -r requirements.txt

langchain
langchain-openai
python-dotenv
psycopg2-binary
sqlalchemy
pgvector
fastapi
uvicorn
#for api's/security
python-jose[cryptography]
passlib[bcrypt]
python-multipart